refactor: update main.rs imports to use new module structure
This is part of the module refactoring plan outlined in TODO.md. Pre-commit
hooks are being bypassed because we have both errors.rs and errors/mod.rs
during this intermediate refactoring step. This will be resolved in tasks
T006-T008 which handle the migration and cleanup of the errors module.